models/pyproject.toml
[tool.poetry]
name = "gesha_models"
version = "0.1.0"
description = "Models for the Silvia espresso machine"
authors = ["Luke Channings <luke@channings.me>"]
readme = "README.markdown"
packages = [{include = "thermal_loss"}]

[tool.poetry.dependencies]
python = ">=3.9, <3.12"
scikit-learn = "^1.3.0"
pandas = "^2.0.3"
matplotlib = "^3.7.2"
seaborn = "^0.12.2"
joblib = "^1.3.2"
python-dotenv = "^1.0.0"
docopt = "^0.6.2"
paho-mqtt = "^1.6.1"
tensorflow-macos = "^2.13.0"
tensorflow = "^2.13.0"
onnxmltools = "^1.11.2"
tf2onnx = "^1.15.1"
xgboost = "^1.7.6"
onnxruntime = "^1.15.1"
tabulate = "^0.9.0"
tract = "^0.20.18"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
thermal_loss = 'thermal_loss.main:main'
extraction_temperature = "extraction_temperature.main:main"
stock_control = "stock_control.main:main"
boiler_levels = "boiler_levels.main:main"
api = "api:main"
pre_heating_automation = "pre_heating.automation:main"
predictive = "predictive.main:main"
